feat: RESTful URLs for Ideas and Tasks side panels
Replace query-param panel URLs (?idea={id}, ?task={id}) with RESTful
paths (/ideas/{id}, /tasks/{id}). Panel open/close/switch updates the
browser URL via History API without page reload, and direct URL access
renders list + panel via SSR.

- Add usePanelUrl hook (pushState/replaceState/popstate)
- Extract IdeasPageContent and TasksPageContent shared server components
- Add /ideas/[ideaUuid] and /tasks/[taskUuid] page routes
- Add middleware HTTP 307 redirects for legacy ?idea= and ?task= URLs
- Update notification and dashboard links to new format
- Remove unused idea-actions.tsx (old standalone detail page)
