+++ src/app/(dashboard)/projects/[uuid]/ideas/page.tsx
@@ -1,177 +1,23 @@
 // src/app/(dashboard)/projects/[uuid]/ideas/page.tsx
-// Server Component - UUID obtained from URL
+// Server Component — list view (no panel selected)
+// Legacy ?idea={id} redirect is handled by middleware (HTTP 307)
 
-import { redirect } from "next/navigation";
-import Link from "next/link";
-import { getTranslations } from "next-intl/server";
-import { Button } from "@/components/ui/button";
-import { Card } from "@/components/ui/card";
-import { Lightbulb } from "lucide-react";
-import { getServerAuthContext } from "@/lib/auth-server";
-import { listIdeas } from "@/services/idea.service";
-import { projectExists } from "@/services/project.service";
-import { checkIdeasAvailability } from "@/services/proposal.service";
-import { batchCommentCounts } from "@/services/comment.service";
-import { IdeaCreateForm } from "./idea-create-form";
-import { IdeasList } from "./ideas-list";
-
-// Filter tab statuses (simplified lifecycle)
-const filterStatuses = ["open", "elaborating", "proposal_created"] as const;
-
-// Status to i18n key mapping
-const statusI18nKeys: Record<string, string> = {
-  open: "open",
-  elaborating: "elaborating",
-  proposal_created: "proposal_created",
-  completed: "completed",
-  closed: "closed",
-};
-
+import { IdeasPageContent } from "./ideas-page-content";
 
 interface PageProps {
   params: Promise<{ uuid: string }>;
-  searchParams: Promise<{ status?: string; assignedToMe?: string; idea?: string }>;
+  searchParams: Promise<{ status?: string; assignedToMe?: string }>;
 }
 
 export default async function IdeasPage({ params, searchParams }: PageProps) {
-  const auth = await getServerAuthContext();
-  if (!auth) {
-    redirect("/login");
-  }
-
   const { uuid: projectUuid } = await params;
-  const { status: filter = "all", assignedToMe, idea: initialIdeaUuid } = await searchParams;
-  const isAssignedToMeFilter = assignedToMe === "true";
-  const t = await getTranslations();
-
-  // Validate project exists
-  const exists = await projectExists(auth.companyUuid, projectUuid);
-  if (!exists) {
-    redirect("/projects");
-  }
-
-  // Get all Ideas (for counting)
-  const { ideas: allIdeas } = await listIdeas({
-    companyUuid: auth.companyUuid,
-    projectUuid,
-    skip: 0,
-    take: 1000,
-  });
-
-  // Get Ideas assigned to me (for counting)
-  const { ideas: myIdeas } = await listIdeas({
-    companyUuid: auth.companyUuid,
-    projectUuid,
-    skip: 0,
-    take: 1000,
-    assignedToMe: true,
-    actorUuid: auth.actorUuid,
-    actorType: auth.type,
-  });
-
-  // Calculate count per status
-  const statusCounts = allIdeas.reduce((acc, idea) => {
-    acc[idea.status] = (acc[idea.status] || 0) + 1;
-    return acc;
-  }, {} as Record<string, number>);
-
-  // Get availability of all Ideas (whether already used by a Proposal)
-  const allIdeaUuids = allIdeas.map(idea => idea.uuid);
-  const availabilityCheck = allIdeaUuids.length > 0
-    ? await checkIdeasAvailability(auth.companyUuid, allIdeaUuids)
-    : { usedIdeas: [] };
-  const usedIdeaUuids = availabilityCheck.usedIdeas.map(u => u.uuid);
-  // idea UUID → proposal UUID mapping
-  const ideaProposalMap: Record<string, string> = {};
-  for (const u of availabilityCheck.usedIdeas) {
-    ideaProposalMap[u.uuid] = u.proposalUuid;
-  }
-
-  // Batch get comment counts
-  const commentCounts = allIdeaUuids.length > 0
-    ? await batchCommentCounts(auth.companyUuid, "idea", allIdeaUuids)
-    : {};
-
-  // Filter by selected status
-  let filteredIdeas = allIdeas;
-
-  // First apply assignedToMe filter if active
-  if (isAssignedToMeFilter) {
-    filteredIdeas = myIdeas;
-  }
-
-  // Then apply status filter if not "all"
-  if (filter !== "all") {
-    filteredIdeas = filteredIdeas.filter((idea) => idea.status === filter);
-  }
+  const { status: filter = "all", assignedToMe } = await searchParams;
 
   return (
-    <div className="p-4 md:p-8">
-      {/* Header */}
-      <div className="mb-6">
-        <h1 className="text-2xl font-semibold text-foreground">{t("ideas.title")}</h1>
-        <p className="mt-1 text-sm text-muted-foreground">
-          {t("ideas.subtitle")}
-        </p>
-      </div>
-
-      {/* Inline Create Form */}
-      <div className="mb-6">
-        <IdeaCreateForm projectUuid={projectUuid} />
-      </div>
-
-      {/* Filter Tabs */}
-      <div className="mb-6 flex gap-2 overflow-x-auto border-b border-border pb-4">
-        <Link href={`/projects/${projectUuid}/ideas`}>
-          <Button variant={filter === "all" && !isAssignedToMeFilter ? "default" : "ghost"} size="sm">
-            {t("ideas.all")} ({allIdeas.length})
-          </Button>
-        </Link>
-        <Link href={`/projects/${projectUuid}/ideas?assignedToMe=true`}>
-          <Button variant={isAssignedToMeFilter && filter === "all" ? "default" : "ghost"} size="sm">
-            {t("ideas.assignedToMe")} ({myIdeas.length})
-          </Button>
-        </Link>
-        {filterStatuses.map((status) => {
-          const count = statusCounts[status] || 0;
-          return (
-            <Link key={status} href={`/projects/${projectUuid}/ideas?status=${status}${isAssignedToMeFilter ? "&assignedToMe=true" : ""}`}>
-              <Button variant={filter === status ? "default" : "ghost"} size="sm">
-                {t(`status.${statusI18nKeys[status]}`)} ({count})
-              </Button>
-            </Link>
-          );
-        })}
-      </div>
-
-      {/* Ideas List */}
-      {filteredIdeas.length === 0 ? (
-        <Card className="flex flex-col items-center justify-center p-12 text-center">
-          <div className="mb-4 flex h-16 w-16 items-center justify-center rounded-full bg-primary/10">
-            <Lightbulb className="h-8 w-8 text-primary" />
-          </div>
-          <h3 className="mb-2 text-lg font-medium text-foreground">
-            {filter === "all" ? t("ideas.noIdeas") : t("ideas.noIdeasWithStatus", { status: t(`status.${statusI18nKeys[filter] || filter}`) })}
-          </h3>
-          <p className="mb-6 max-w-sm text-sm text-muted-foreground">
-            {filter === "all"
-              ? t("ideas.startByAdding")
-              : t("ideas.ideasWithStatus")}
-          </p>
-        </Card>
-      ) : (
-        <IdeasList
-          ideas={filteredIdeas.map(idea => ({
-            ...idea,
-            commentCount: commentCounts[idea.uuid] || 0,
-          }))}
-          projectUuid={projectUuid}
-          currentUserUuid={auth.actorUuid}
-          usedIdeaUuids={usedIdeaUuids}
-          ideaProposalMap={ideaProposalMap}
-          initialSelectedIdeaUuid={initialIdeaUuid}
-        />
-      )}
-    </div>
+    <IdeasPageContent
+      projectUuid={projectUuid}
+      filter={filter}
+      isAssignedToMeFilter={assignedToMe === "true"}
+    />
   );
 }
